Moved stores, commenting, tidying, update for Simple Injector, added support for non-store managed configuration.
