--- Mamesaver/Settings/GeneralSettingsStore.cs
@@ -1,9 +1,0 @@
-﻿using Mamesaver.Models.Settings;
-
-namespace Mamesaver.Settings
-{
-    public class GeneralSettingsStore : SettingsStore<GeneralSettings>
-    {
-        public override string Filename => "settings.xml";
-    }
-}
